requirements.txt
google-generativeai
requests
feedparser
beautifulsoup4
googlenewsdecoder
